Change the advertised topic from actuator_controls_0 to manual_control_setpoint
With this commit the module publishes to the manual_control_setpoint topic instead of actuator_controls_0. This architecture is closer to simulating an RC since the attitude control processes the input like it was an RC input.

This is the first commit for which the module works in SITL.

On an FMU the model crashes. It's not yet clear why this happens.